extract detail view and font
