+++ packages/widget/src/index.ts
@@ -1,3 +1,9 @@
 // All files which will be accessible when the widget is installed via npm are declared here
 export * from './docmaps-widget';
 export * from './docmap-controller';
+export { emptyMetadataMessage } from './detail-view.ts';
+export { createGridItem } from './detail-view.ts';
+export { createMetadataGrid } from './detail-view.ts';
+export { renderDetailsView } from './detail-view.ts';
+export { addLinkToDocumentHeader } from './font.ts';
+export { loadFont } from './font.ts';
